requirements/testing.txt
# Testing dependencies
# --------------------
pytest-cov>=4,<5